# openapi.yaml
openapi: 3.1.0
info:
  title: Image Bridge API
  version: 1.0.0
  description: |
    API for generating images via AI and saving them to a specified path.
    Designed for integration with Custom GPT Actions.
servers:
  - url: https://gpt-image-bridge.onrender.com
    description: Production server (Render)
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check endpoint
      description: Returns 200 OK if the service is running
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /generate-image:
    post:
      operationId: generateImage
      summary: Generate an image asset and save it to the requested path
      description: |
        Generates an image based on the provided prompt and saves it to the specified path.
        Requires either OAuth2 or API Key authentication.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                prompt:
                  type: string
                  minLength: 1
                  maxLength: 2000
                  description: Detailed text description of the image to generate
                out_path:
                  type: string
                  minLength: 1
                  pattern: "^[a-zA-Z0-9_\\-\\./]+\\.(png|jpg|jpeg|webp)$"
                  description: |
                    The filename or relative subpath where the image should be saved. 
                    Must NOT start with a slash or contain '..'.
                    Example: 'assets/hero.png' or 'image.png'.
              required: [prompt, out_path]
      responses:
        '200':
          description: Image generation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateImageResponse'
        '400':
          description: Bad request - invalid input parameters or path traversal attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during image generation or storage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /generate-images:
    post:
      operationId: batchGenerateImages
      summary: Generate multiple image assets in a single request
      description: |
        Generates multiple images based on provided prompts. 
        Prompts are prefixed with an optional style anchor for consistency.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGenerateImageRequest'
      responses:
        '200':
          description: Batch generation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGenerateImageResponse'
        '400':
          description: Bad request - invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during batch processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    GenerateImageResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          description: Whether the operation was successful
        path:
          type: string
          description: The requested relative output path
        absolute_path:
          type: string
          description: The resolved absolute path where image was saved
        mime_type:
          type: string
          description: MIME type of the generated image (e.g., image/png)
        width:
          type: integer
          nullable: true
          description: Width of the generated image in pixels
        height:
          type: integer
          nullable: true
          description: Height of the generated image in pixels
        message:
          type: string
          description: Status message
      required: [ok, path, absolute_path, mime_type, message]
    
    BatchGenerateImageRequest:
      type: object
      additionalProperties: false
      properties:
        style_anchor:
          type: string
          maxLength: 1000
          description: Shared style description prepended to each item prompt
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/GenerateImageItem'
      required: [items]

    GenerateImageItem:
      type: object
      additionalProperties: false
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
          description: Text description of the image
        out_path:
          type: string
          minLength: 1
          pattern: "^[a-zA-Z0-9_\\-\\./]+\\.(png|jpg|jpeg|webp)$"
          description: Relative save path
      required: [prompt, out_path]

    BatchGenerateImageResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/GenerateImageResponse'
        message:
          type: string
      required: [ok, results, message]

    
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
          description: A high-level error code or summary
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required: [ok, error]
    
    ErrorDetail:
      type: object
      additionalProperties: false
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
      required: [msg, type]

  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Recommended for Custom GPT Actions.
        Use the key found in your Render dashboard environment settings.